+++ src/Serilog/Settings/KeyValuePairs/ISettingsSourceBuilder.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2013-2017 Serilog Contributors
+// Copyright 2013-2017 Serilog Contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Serilog.Settings
+namespace Serilog.Settings.KeyValuePairs
 {
     /// <summary>
     /// A builder that allows to combine sources of key-value settings in a fluent way.
